Create Week6
Create a function that takes a text file as input and returns a vector of size 26 with the frequency in percent of
each character a, b, . . . z (not sensitive to case.) The frequency must be computed as the number of occurences
divided by the total number of characters from a to z that occur, multiplied by one hundred. All other letters
such as ø and ä as well as all symbols must be ignored.