Create Week6-LanguageDet
Different languages use the different letters with different frequencies and this can be used to determine in
which language a text is written. In this exercise you should use the function created in the previous exercise
to compute the frequencies of letters in a text. Given this vector of frequencies you can compute the squared
error between the frequencies in the text and the (average) frequencies in the language.

The formula is given by     E = np.sum((languageFrequencies - freq)**2, axis=1)
where Fit is the frequency of letter i in the text and Fil is the frequency of letter i in the language.
The language which has the lowest squared error is the one that best matches the text in terms of the letter frequency.